requirements.txt
beautifulsoup4>=4.9.3
questionary=>1.10.0
PyYAML>=5.4.1
